chore(icons): clean up and improve consistency
