+++ src/common-prop-types/index.js
@@ -15,3 +15,7 @@ export const sizePropType = mutuallyExclusive(
     ['small', 'large'],
     propTypes.bool
 )
+
+export const iconPropType = {
+    className: propTypes.string,
+}
