+++ src/icons/Arrow.js
@@ -1,40 +1,31 @@
 import React from 'react'
-import css from 'styled-jsx/css'
+import { iconPropType } from '../common-prop-types'
+import { styles } from './styles.js'
 
-export const styles = css`
-    svg {
-        fill: inherit;
-        height: 24px;
-        width: 24px;
-        vertical-align: middle;
-        pointer-events: none;
-    }
-`
-
-export function ArrowDown() {
+export function ArrowDown({ className }) {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
-            width="48"
-            height="48"
             viewBox="0 0 48 48"
+            className={className}
         >
             <path d="M14 20l10 10 10-10z" />
             <style jsx>{styles}</style>
         </svg>
     )
 }
+ArrowDown.propTypes = iconPropType
 
-export function ArrowUp() {
+export function ArrowUp({ className }) {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
-            width="48"
-            height="48"
             viewBox="0 0 48 48"
+            className={className}
         >
             <path d="M14 28l10-10 10 10z" />
             <style jsx>{styles}</style>
         </svg>
     )
 }
+ArrowUp.propTypes = iconPropType
