+++ src/icons/AttachFile.js
@@ -1,0 +1,18 @@
+import React from 'react'
+import { iconPropType } from '../common-prop-types'
+import { styles } from './styles.js'
+
+export function AttachFile({ className }) {
+    return (
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            viewBox="0 0 48 48"
+            className={className}
+        >
+            <path d="M33 12v23c0 4.42-3.58 8-8 8s-8-3.58-8-8V10c0-2.76 2.24-5 5-5s5 2.24 5 5v21c0 1.1-.89 2-2 2-1.11 0-2-.9-2-2V12h-3v19c0 2.76 2.24 5 5 5s5-2.24 5-5V10c0-4.42-3.58-8-8-8s-8 3.58-8 8v25c0 6.08 4.93 11 11 11s11-4.92 11-11V12h-3z" />
+            <style jsx>{styles}</style>
+        </svg>
+    )
+}
+
+AttachFile.propTypes = iconPropType
