+++ src/icons/Status.js
@@ -1,7 +1,9 @@
-import propTypes from 'prop-types'
 import React from 'react'
+import propTypes from 'prop-types'
 
 import { theme, spacers } from '../theme.js'
+import { iconPropType } from '../common-prop-types'
+import { styles } from './styles.js'
 
 /**
  * @param {Object} props
@@ -12,27 +14,15 @@ export function Valid({ className }) {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
-            width="48"
-            height="48"
             viewBox="0 0 48 48"
             className={className}
         >
             <path d="M24 4C12.95 4 4 12.95 4 24c0 11.04 8.95 20 20 20 11.04 0 20-8.96 20-20 0-11.05-8.96-20-20-20zm-4 30L10 24l2.83-2.83L20 28.34l15.17-15.17L38 16 20 34z" />
-            <style jsx>{`
-                svg {
-                    fill: ${theme.valid};
-                    width: 24px;
-                    height: 24px;
-                    margin-right: ${spacers.dp4};
-                }
-            `}</style>
+            <style jsx>{styles}</style>
         </svg>
     )
 }
-
-Valid.propTypes = {
-    className: propTypes.string,
-}
+Valid.propTypes = iconPropType
 
 /**
  * @param {Object} props
@@ -43,27 +33,15 @@ export function Warning({ className }) {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
-            width="48"
-            height="48"
             viewBox="0 0 48 48"
             className={className}
         >
             <path d="M2 42h44L24 4 2 42zm24-6h-4v-4h4v4zm0-8h-4v-8h4v8z" />
-            <style jsx>{`
-                svg {
-                    fill: ${theme.warning};
-                    width: 24px;
-                    height: 24px;
-                    margin-right: ${spacers.dp4};
-                }
-            `}</style>
+            <style jsx>{styles}</style>
         </svg>
     )
 }
-
-Warning.propTypes = {
-    className: propTypes.string,
-}
+Warning.propTypes = iconPropType
 
 /**
  * @param {Object} props
@@ -74,23 +52,15 @@ export function Error({ className }) {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
-            width="48"
-            height="48"
             viewBox="0 0 48 48"
             className={className}
         >
             <path d="M24 4C12.96 4 4 12.95 4 24s8.96 20 20 20 20-8.95 20-20S35.04 4 24 4zm2 30h-4v-4h4v4zm0-8h-4V14h4v12z" />
-            <style jsx>{`
-                svg {
-                    fill: ${theme.error};
-                    width: 24px;
-                    height: 24px;
-                    margin-right: ${spacers.dp4};
-                }
-            `}</style>
+            <style jsx>{styles}</style>
         </svg>
     )
 }
+Error.propTypes = iconPropType
 
 /**
  * @param {Object} props
@@ -101,27 +71,15 @@ export function Info({ className }) {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
-            width="48"
-            height="48"
             viewBox="0 0 48 48"
             className={className}
         >
             <path d="M24 4C12.95 4 4 12.95 4 24s8.95 20 20 20 20-8.95 20-20S35.05 4 24 4zm2 30h-4V22h4v12zm0-16h-4v-4h4v4z" />
-            <style jsx>{`
-                svg {
-                    fill: ${theme.info};
-                    width: 24px;
-                    height: 24px;
-                    margin-right: 4px;
-                }
-            `}</style>
+            <style jsx>{styles}</style>
         </svg>
     )
 }
-
-Error.propTypes = {
-    className: propTypes.string,
-}
+Info.propTypes = iconPropType
 
 /**
  * @param {Object} props
@@ -132,9 +90,7 @@ export function Loading({ className }) {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
-            width="48"
-            height="48"
-            viewBox="22 22 44 44"
+            viewBox="0 0 48 48"
             className={className}
         >
             <circle
@@ -186,10 +142,7 @@ export function Loading({ className }) {
         </svg>
     )
 }
-
-Loading.propTypes = {
-    className: propTypes.string,
-}
+Loading.propTypes = iconPropType
 
 /**
  * @param {Object} props
