+++ src/icons/styles.js
@@ -1,0 +1,11 @@
+import css from 'styled-jsx/css'
+
+export const styles = css`
+    svg {
+        fill: inherit;
+        height: 24px;
+        width: 24px;
+        vertical-align: middle;
+        pointer-events: none;
+    }
+`
